+++ tcvectordb/rpc_base_collection.go
@@ -405,6 +405,8 @@ func (r *rpcImplementerCollection) toCollection(collectionItem *olama.CreateColl
 					vector.Params = &IVFPQParams{M: index.Params.M, NList: index.Params.Nlist}
 				case IVF_SQ4, IVF_SQ8, IVF_SQ16:
 					vector.Params = &IVFSQParams{NList: index.Params.Nlist}
+				case IVF_RABITQ:
+					vector.Params = &IVFRabitQParams{NList: index.Params.Nlist, Bits: index.Params.Bits}
 				}
 			}
 			coll.Indexes.VectorIndex = append(coll.Indexes.VectorIndex, vector)
@@ -483,6 +485,11 @@ func optionRpcParamsFromIndexParams(column *olama.IndexColumn, v IndexParams) {
 			column.Params.M = param.M
 			column.Params.Nlist = param.NList
 		}
+	case *IVFRabitQParams:
+		if param != nil {
+			column.Params.Nlist = param.NList
+			column.Params.Bits = param.Bits
+		}
 	default:
 		log.Printf("[Warning] unknown type: %v", reflect.TypeOf(v))
 	}
@@ -508,6 +515,11 @@ func optionRpcParams(column *olama.IndexColumn, v VectorIndex) {
 		if param, ok := v.Params.(*IVFPQParams); ok && param != nil {
 			column.Params.M = param.M
 			column.Params.Nlist = param.NList
+		}
+	case IVF_RABITQ:
+		if param, ok := v.Params.(*IVFRabitQParams); ok && param != nil {
+			column.Params.Nlist = param.NList
+			column.Params.Bits = param.Bits
 		}
 	}
 }